rattachement/src/main/webapp/app/entities/habilitation/habilitation/route/habilitation-routing-resolve.service.spec.ts
import { TestBed } from '@angular/core/testing';
import { HttpResponse } from '@angular/common/http';
import { HttpClientTestingModule } from '@angular/common/http/testing';
import { ActivatedRouteSnapshot, ActivatedRoute, Router, convertToParamMap } from '@angular/router';
import { RouterTestingModule } from '@angular/router/testing';
import { of } from 'rxjs';

import { IHabilitation, Habilitation } from '../habilitation.model';
import { HabilitationService } from '../service/habilitation.service';

import { HabilitationRoutingResolveService } from './habilitation-routing-resolve.service';

describe('Habilitation routing resolve service', () => {
  let mockRouter: Router;
  let mockActivatedRouteSnapshot: ActivatedRouteSnapshot;
  let routingResolveService: HabilitationRoutingResolveService;
  let service: HabilitationService;
  let resultHabilitation: IHabilitation | undefined;

  beforeEach(() => {
    TestBed.configureTestingModule({
      imports: [HttpClientTestingModule, RouterTestingModule.withRoutes([])],
      providers: [
        {
          provide: ActivatedRoute,
          useValue: {
            snapshot: {
              paramMap: convertToParamMap({}),
            },
          },
        },
      ],
    });
    mockRouter = TestBed.inject(Router);
    jest.spyOn(mockRouter, 'navigate').mockImplementation(() => Promise.resolve(true));
    mockActivatedRouteSnapshot = TestBed.inject(ActivatedRoute).snapshot;
    routingResolveService = TestBed.inject(HabilitationRoutingResolveService);
    service = TestBed.inject(HabilitationService);
    resultHabilitation = undefined;
  });

  describe('resolve', () => {
    it('should return IHabilitation returned by find', () => {
      // GIVEN
      service.find = jest.fn(id => of(new HttpResponse({ body: { id } })));
      mockActivatedRouteSnapshot.params = { id: 123 };

      // WHEN
      routingResolveService.resolve(mockActivatedRouteSnapshot).subscribe(result => {
        resultHabilitation = result;
      });

      // THEN
      expect(service.find).toBeCalledWith(123);
      expect(resultHabilitation).toEqual({ id: 123 });
    });

    it('should return new IHabilitation if id is not provided', () => {
      // GIVEN
      service.find = jest.fn();
      mockActivatedRouteSnapshot.params = {};

      // WHEN
      routingResolveService.resolve(mockActivatedRouteSnapshot).subscribe(result => {
        resultHabilitation = result;
      });

      // THEN
      expect(service.find).not.toBeCalled();
      expect(resultHabilitation).toEqual(new Habilitation());
    });

    it('should route to 404 page if data not found in server', () => {
      // GIVEN
      jest.spyOn(service, 'find').mockReturnValue(of(new HttpResponse({ body: null as unknown as Habilitation })));
      mockActivatedRouteSnapshot.params = { id: 123 };

      // WHEN
      routingResolveService.resolve(mockActivatedRouteSnapshot).subscribe(result => {
        resultHabilitation = result;
      });

      // THEN
      expect(service.find).toBeCalledWith(123);
      expect(resultHabilitation).toEqual(undefined);
      expect(mockRouter.navigate).toHaveBeenCalledWith(['404']);
    });
  });
});
